chore: migrate example dapp to src/ dir
