+++ packages/example/src/pages/api/auth/[...nextauth].ts
