+++ packages/example/src/pages/global.css
