+++ packages/example/src/types/next-auth.d.ts
